Handle exception raised by incorrect access to editor
There appears to be a reentrancy case in which a blur causes
the lost focus of the HTML editor to be reentrant. In a normal
circumstance, this causes an error to be raised and the code to
exit, with the error being reported. This change detects that
condition and exits, but doesn't raise an exception.

Change-Id: If083b318431832ed6314bd05b7e4ea171d7b4178