[AOM] Element Reflection for the Accessibility Pane
This is a small change for supporting aria relationships set via the IDL
interface on element. This requires special casing because we can't rely
on the `value of the content attribute, as depending on the scope of the
set elements, the content attribute may be set to the empty string. In
this case we can just also check if any related nodes were supplied via
the protocol and construct the list from these.

Spec:https://whatpr.org/html/3917/common-dom-interfaces.html#reflecting-content-attributes-in-idl-attributes:element

Bug:981423
Change-Id: I81130ae5be94f74c0ba82b9f6709011b137f610f