Datagrid: Update Roll to Application and Add A11y Text Methods
**Rough First draft**

This CL updates the Datagrid so that it has the role of application.
This causes screen readers like NVDA to interact with it in an
application mode, similar to focus mode. It also adds the ability
for each node to specify how it wants it's accessible text read, and
updates the accessible text for the datagrid on each selection or
on expanding/collapsing a node

Change-Id: I7c563fbdb358e3f68e62bac1c3fefcddb8f65307