Split ModelData.uiLocationToRawLocations into two methods
This CL prepares ModelData.uiLocationToRawLocations for asyncification.
To prune the call-tree we need to asyncify, we split the
ModelData.uiLocationToRawLocations into an async and an non-async part.

Since we know the exact code path the call-site in {SourceFormatter}
is going to take, we can build a custom variation of
{uiLocationToRawLocations} that won't require any async parts. This
way, we do not need to asyncyfiy the SourceMapping interface, all
implementations and all use-sites.

R=sigurds@chromium.org

Bug: chromium:1032016, chromium:1032920
Change-Id: I048222d470ce8772d0714d28894b3797952b3519