Restore sorted object property order in expanded view
A long time ago in a galaxy far, far away, there was a property order
mismatch between the DevTools Console Preview and the expanded view.

For example:

       input> ({ c: 1, b: 1, a: 1 })
     preview> {c: 1, b: 1, a: 1}
    expanded> {a: 1, b: 1, c: 1, __proto__: Object}

We then attempted [1] to remove the confusing mismatch as follows:

       input> ({ c: 1, b: 1, a: 1 })
     preview> {c: 1, b: 1, a: 1}
    expanded> {c: 1, b: 1, a: 1, __proto__: Object}

In addition to avoiding the mismatch, this patch made DevTools more
closely resemble the internal shapes of objects [2] instead of hiding
this information from developers.

However, since shipping, we've heard overwhelming feedback in favor of
restoring the old behavior, where the expanded view is sorted. As a
result of this feedback, this patch restores the sorting logic.

[1]: https://chromium-review.googlesource.com/c/chromium/src/+/1806457
[2]: https://mathiasbynens.be/notes/shapes-ics

Bug: chromium:989514, chromium:1008910
Change-Id: I5438987d73b8b980080156b5dd8b11f3f19fcc10