Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Adding Kayce as reviewer who has been maintaining this panel.
> 
> I did notice in the gif that the Youtube video is no longer tabbable. Was that change intended? I think we still want users to be able to tab and go to the Youtube video for more information.

Indeed - as it turns out the YouTube video is not a YouTube video at all, but instead a static image that hyperlinks to the release notes. Interacting with it (click or keyboard) would performs the same action as pressing the "Learn More" button.

With the image link redundant, in my mind it felt more natural to have tab order end on the "Close" button. Though, I don't feel strongly about this change; I can see a desire to keep the keep the image tab accessible if we want to make it link to someplace different.

Patch-set: 2
Reviewer: Gerrit User 1363999 <1363999@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>