Update patch set 6
Patch Set 6:

> Patch Set 2:
> 
> This looks good! It's great that the SR doesn't just dump the entire contents of the table anymore 😊
> 
> Just a couple things to consider:
> - nit: Every row reads out "Coverage List" first, which is a bit redundant
Good point. I've updated to now only read out the grid name if the grid is receiving focus (for context), otherwise the label just updates to the next row
> - The dataGrid is labeled as "Coverage List" but it's really a table/grid.
Updated
> - As-is, a user can't press [T] to navigate to next table with NVDA. Once on the grid, they can use [Insert]+[Space] to toggle out of application mode and navigate the table as usual, so it's not completely blocking, but you might want to check with other a11y folks to see how impactful that is.
That is true. So, marking the grid as an application means that screen readers no longer recognizes the tables with single key navigation by default. 

The goal of this change is that SR users wouldn't have to toggle modes or get the massive content dump when hitting a datagrid. Since grid is a Div, a div recieving focus is always going to dump all text unless role=applicaiton is set. And, since it isn't really a standard table or grid, it's more of a custom interaction - I feel that application is a reasonable approach.

I've updated the default aria-label to mention that you can toggle to navigate cell by cell, but my goal is this wouldn't be necessary for SR users any more.

For more info:
https://github.com/nvaccess/nvda/issues/7360
http://blog.jantrid.net/2015/12/woe-aria-surprisingly-but-ridiculously.html
https://github.com/nvaccess/nvda/issues/7807
https://www.davidmacd.com/blog/does-aria-label-override-static-text.html

Patch-set: 6