Datagrid: Update Role to Application and Add aria-live
**Test CL, DO NOT MERGE**

This tests if aria-live works better than aria-label with VO. If successful
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1967346
will be updated to use this approach.

Original description

This CL updates the Datagrid so that it has the role of application.
This causes screen readers like NVDA to interact with it in an
application mode, similar to focus mode. It also adds the ability
for each node to specify how it wants its accessible text read, and
updates the accessible text for the datagrid on each selection or
on expanding/collapsing a node.

The goal of this change is that SR users wouldn't have to toggle modes
or get the massive content dump when hitting a datagrid.
Since grid is a div, a div receiving focus is expected by SR's to dump all
text unless role=applicaiton is set, which is not desirable for datagrids.
It is the recommended approach
for custom div sections that require specific user interaction
to mark role="application" and specify through custom labels how SR
users should interact. https://github.com/nvaccess/nvda/issues/7360
This is the same approach used by google sheets and excel online.

Additionally, SRs that support changing modes  can still navigate
the table in browse mode by changing modes after the
datagrid receives focus.

More references:
http://blog.jantrid.net/2015/12/woe-aria-surprisingly-but-ridiculously.html
https://github.com/nvaccess/nvda/issues/7807
https://www.davidmacd.com/blog/does-aria-label-override-static-text.html

Change-Id: I8517ed874d5678ad7136a0d89897d195838aeddb