Animations: make the full node text visible when keyboard focused.
This CL updates the linkified DOM node links in the animations tool to
have a scroll bar when focused by the keyboard.

Previously, the full text of the link could not be read by keyboard users,
who did not have access to the tooltip.
Now, when a row is focused, if it extends beyond
the first column, a scrollbar is added to allow keyboard users to see
the full text.

Before: https://imgur.com/woBABMz

After: https://imgur.com/l2wLZFy

Bug: 963183
Change-Id: I25a3f5e0da0eefeae87f72ea109338b66aed766a