Remove restriction to eager (popover) evaluation.
This undoes http://chromium-review.googlesource.com/c/chromium/src/+/1835357,
which limited the eager popover evaluation to the innermost scope, which
is the only scope for which we can produce reliable results right now.
For the outer scopes, it might still produce the right results, but we
can't guarantee that. But the feature is still very useful and so we are
restoring the behavior here, and will follow up with a proper fix and UX
later.

Bug: chromium:995036, chromium:1033022
Change-Id: Ie4716f4fce11fdb092cc6b884f69820984f5960a