Coverage: Update URL cell in datagrid screen reader label to full url
This cl updates the URL cell of the datagrid in the coverage tool
to use the full url as the accessible text when using browse mode
to view the cell.

Previously, if a long url was truncated, the screen reader would read out
the individual truncated divs. Now, it reads out the full url when navigating
to the cell.

Before: https://imgur.com/59etRQ8

After: https://imgur.com/UUKLG1b

Bug: 963183
Change-Id: I698cda0596080e62ff68ad54f8215f56ee74dcdb