Audits: give focus to New Audit button after completing a report.
Previously, when completing a report, the focus for the Audits pane
would be completely lost. This CL sets focus to the new audit button in
the toolbar after completing an audit, and sets data-keyboard-focus based
on whether or not the start audit button had data-keyboard-focus set
before starting the audit.

Before:

After:

Bug: 963183
Change-Id: Ib1a98a53e942c9a1cd11b5d64ad9cbe64842e019