Update patch set 4
Patch Set 4:

Tested locally, behavior looks good.

One concern with this change is that it's possible to shrink the sidebar until it disappears. Previously, after opening and closing DevTools, it would pop open back to the default height, but this way it stays hidden. It would be nice if a minimum height of the sidebar could be enforced so users don't "lose" the panel.

Patch-set: 4
Reviewer: Gerrit User 1363999 <1363999@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>