Fix possible Null Pointer Exception with data grid background color
This tiny CL adds a null check on this.dataGrid of the Network Data
Grid Node, which is an optional (nullable) property, in the
backgroundColor() method, to prevent a Null Pointer Exception when
this.dataGrid is set to null.

We've received a significant number of automated reports from users
that a Null Pointer Exception is being thrown during some occurrences
of the network data grid refresh, likely caused by some race condition.

As a fallback we can presume the node is not selected if it is also
not attached to any dataGrid, which is obviously preferrable to an
uncaught exception.

Change-Id: I90d16df0dcacc0df874af37bcb63351be3e667f6