Fix possible Null Pointer Exception with data grid background color
This tiny CL adds a null check on this.dataGrid of the Network Data
Grid Node, which is an optional (nullable) property, in the
backgroundColor() method, to prevent a Null Pointer Exception when
this.dataGrid is set to null.

During some occurrences of the Network Data Grid refresh, there may be
a race condition where this.dataGrid is set to null.

As a fallback we can presume the node is not selected if it is also
not attached to any dataGrid, which is obviously preferable to an
uncaught exception.

Change-Id: I90d16df0dcacc0df874af37bcb63351be3e667f6