(WIP) Infobar Improvements
This CL adds a new action button to the Infobar component so users may
take immediate action from an Infobar.

The Infobar "Pretty-print this minified file?" has been augmented to
use this feature so users can click on a button in the Infobar itself
to pretty-print the code (rather than having to click "more..." and
then follow the instructions in the Infobar).

This CL is a work-in-progress.

Change-Id: Id2e695af367fb8f5a1c489fa8d3d94c8974458fb