Profiler: update start/stop to retain focus on press
This CL updates both the start/stop button and the record icon to
retain focus when pressed to start a profile for the memory or
JavaScript Profiler tool.

As part of starting a recording, the SDK ensures there are no other
active profiles, and this triggers logic to disable the start button, causing
focus to be lost before the button is re-enabled to allow users to stop
recording. This CL adds a focus restorer to give focus back to the button
after completing the toggle and changing the button text.

Before:

After:

Bug: 963183
Change-Id: Idbbc4468ec738e0db16e44f22b89c9fa7c032db8