Performance: remove focus on Overview Grid resizers when grid is disabled
This CL fixes a small issue that allowed users to focus on the overview
grid resizers before the overview grid was enabled. This caused
issues in the Performance panel when keyboard navigating before starting
a trace: focus would appear to be lost for 2 tab stops between the toolbar
and the Landing page.

Change is also evident when tabbing through network panel when not
recording and log is cleared (overview grid is also disabled in this case)

Before: https://imgur.com/fudKAcn

After: https://imgur.com/G4wujIL

Bug: 963183
Change-Id: I3efc67fa5b9e4a28fc589ae67d19f7305745d4c7