Squelch CSS race condition error on edits
There are minor race conditions in which the StylePropertyTreeElement
is tracking its own CSSProperty and the property is deleted from the
CSSStyleDeclaration to which it belongs. The conditions are particularly
difficult to reproduce, and the user experience does not seem to raise
an issue, probably because changes to the underlying SDK model cause
immediate refreshes of the entire tree).

This change modifies the behavior to not cause an error in the event
that the race condition is encountered by treating edits as having the
entire string, so that user edits aren't dropped.

Change-Id: I2777196cb88fa2928f8a5471398b0a58c8c67853