Silent failure in _innerStyleSheetChanged
An unhandled exception occurs periodically in the ResourceMapping#
_innerStyleSheetChanged function. This is because of at least two
cases which fire from the SDK, one of which is in the case where
CSS changes from the renderer.

The bug appears to derive from Closure compiler's allowing of a
mis-typing to occur in the SDK area due to inferred default-values.
Starting in CSSModel.js:

    _fireStyleSheetChanged(styleSheetId, edit) {

This function has `edit` defined as {!Edit=} but, if a value isn't
actually passed in (which is the case in at least two instances),
no value is passed and the `edit` property of the event data is
explicitly initialized to `undefined`.

That subsequently calls into ResourceMapping.js _styleSheetChanged,
an event handler, which reaches into `event.data.edit` which is
assumed to be non-null.

That is further propagated into the _styleSheetChanged implementation
which accepts a `{!SDK.CSSModel.Edit}`, and Closure thinks that's valid
but it in fact isn't. Finally, this is retrieved from a type-inferred
object in this._edits, but the type inference is pulled from the above
invalid claim.

Change-Id: I94be194eff677308a4ff93833508d3e93b501f94