Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm not a fan of this change.
> 
> They say that naming is one of the only two Hard Problems of computer science. When we limit column-length to 80, this exacerbates problems with naming by encouraging them to be artificially short. (Note: I'm not advocating that we have particularly long names. What I'm saying is, if we say "we're going to limit names to 80," that is going to either limit the length of names or make lines which are > 80 columns to be more of a nuisance to check in against. This is aggravated by the fact that we generally have wide-screen monitors, rendering a need for such narrow documents moot.
> 
> I'm not going to -1 this change, but I'd like to put this out there that I am very much opposed to it.

It is the Chromium style guide to have a column length of 80 characters. I acknowledge concerns with the limit, but in that case I think it is better to resolve that discussion in a Chromium-wide discussion.

This CL merely brings DevTools inline with the wider Chromium policy.

Patch-set: 1
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0
Label: Auto-Submit=0