Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Can you give an example of interacting with DevTools that causes this condition? I would have thought typing something like `Object.create(null).` into the console, but it doesn't look like it. 
> Thanks!

Unfortunately I do not have a repro, as I mention in the description the callFunctionJSON needs to fail due to a protocol error.

This is clearly possible from looking at the code and a type checker would enforce this change which I think is a good enough justification for the change.

Patch-set: 2
Reviewer: Gerrit User 1341668 <1341668@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>