Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > I'm not a fan of this change.
> > > > 
> > > > They say that naming is one of the only two Hard Problems of computer science. When we limit column-length to 80, this exacerbates problems with naming by encouraging them to be artificially short. (Note: I'm not advocating that we have particularly long names. What I'm saying is, if we say "we're going to limit names to 80," that is going to either limit the length of names or make lines which are > 80 columns to be more of a nuisance to check in against. This is aggravated by the fact that we generally have wide-screen monitors, rendering a need for such narrow documents moot.
> > > > 
> > > > I'm not going to -1 this change, but I'd like to put this out there that I am very much opposed to it.
> > > 
> > > It is the Chromium style guide to have a column length of 80 characters. I acknowledge concerns with the limit, but in that case I think it is better to resolve that discussion in a Chromium-wide discussion.
> > > 
> > > This CL merely brings DevTools inline with the wider Chromium policy.
> > 
> > I understand that. But we're now in a separate repo. I don't find a burning need for us to migrate to adopt the wider Chromium policy, particularly now that we're in a separate repo.
> 
> I'm also not a fan of narrow column widths. As we're now a separate repository, do all of the same standards still apply? 120 was chosen at some point and I'm not sure why we need to move to 80 now. Maybe we can simply call this out as a refinement to the standard for our project?

Adding my 2 cents - I also think that introducing an 80 character limit will hurt readability, particularly because this project did not previously enforce the limit. With a new, shorter limit, we will end up with newer parts of the reop that look significantly different than older parts. In the section of the Google style guide that discusses the 80 character limit (https://google.github.io/styleguide/cppguide.html#Line_Length) it says "We recognize that this rule is controversial, but so much existing code already adheres to it, and we feel that consistency is important.". I think that changing the limit to conflict with our project's existing code will introduce a project-level inconsistency which goes against Chromium's reasoning for enforcing an 80 character limit in the first place.

Patch-set: 1
CC: Gerrit User 1330301 <1330301@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>