Update patch set 9
Patch Set 9:

(1 comment)

> Patch Set 9:
> 
> (1 comment)
> 
> A few observations:
> 
>  - When focus is on the "Call Stack" header, pressing Up Arrow or Down Arrow will cause scrolling to occur. Arguably, at the very least, Down Arrow should cause the first call stack item to be selected.
>  - The mouse-over highlight wins against the selected highlight. That is, focus on the 1st call stack frame, put the mouse over the 3rd frame, and arrow-down to the 4th frame. While focus is on the 3rd frame, you'll see that no change in coloration occurs, and it looks as if focus has disappeared.
>  - Right-click "Restart frame" causes focus to be lost from the Call Stack window
> 
> Other than that, there's a nit in the CR.

We're treating the headers as buttons, so it makes sense to me that the arrow keys would scroll like they do when focus is on a native button.

I think restart frame moving focus makes sense as well, as it sends you to the point in the code where that frame is.

The hover issue should be fixed now.

Patch-set: 9