Update patch set 5
Patch Set 5:

Changed code to avoid repeating unnecessary code.

I'm don't entirely love taking the text from an element's textContent, but it is currently the easiest way to access that information without additional overhead.  Let me know if I should approach this differently or if this is okay.

Patch-set: 5
Work-in-progress: false