Makes prototype functions configurable
In some tests we need to import functionality from utilities.js, which
hijacks the prototype of several built-ins, in some cases with
Object.defineProperty. If defineProperty is called more than once, and
the property is not configurable, the call will fail. This CL updates
the defineProperty calls such that the properties set are configurable,
which means they can be set multiple times.

Change-Id: If5d4a11387279bbc6f203db2a7957e98b61e25ce