Update patch set 15
Patch Set 15:

Hey, so sorry this is taking a long time for us to review. As you've noticed we're making a lot of changes to the architecture of, and infrastructure supporting, the DevTools frontend. In honesty it may be some time before we have capacity to help with debugging as our top priority is around improving testability and test coverage. Happy to chat more, so please do ping me if that would be helpful.

Patch-set: 15
Reviewer: Gerrit User 1189020 <1189020@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>