Defensive null check in StylesSidebarPane
Error reports indicate that there are occasional non-fatal errors
related to attempting to call the 'addRule()' function against
a null reference. Investigating, I found that there are three cases
in which Closure is not detecting that CSSModel#addRule was being
called without checking whether the CSSModel reference was null,
because it was not honoring the nullability declaration of
ElementsSidebarPane#cssModel() (which is declared to be of type
{?SDK.CSSModel}).

There are actually a number of other cases in this file which check
for this condition, which suggests to me that it's not actually
an expected invariant.

Change-Id: I287e22490836cc1194bee218f62e2474cefe80d1