Update patch set 1
Patch Set 1:

The row-title changes look good and work well. However, the horizontal scroll bar is sometimes hidden until scrolling down to see it, so it's not obvious that the view can be scrolled horizontally. You can repro this with the "Receive Response" and "Receive Data" event types as they have a lot of vertical information.

Maybe move the overflow rules into the parent container?

Patch-set: 1
CC: Gerrit User 1363999 <1363999@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>