Fix file references to wasm parser
In
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1941748
a new wasmparser worker was introduced, which uses files from a
third_party file. However, the source files of the third_party folder
were added to `all_devtools_files` instead of `all_devtools_modules`.
This causes the Chromium build to not bundle these files in, as they
were not in the GRD file.

We will fix the `check_gn` script to properly error on this issue.

Fixed: 1041415
Change-Id: Iae077c91ac62ca794086229d7b73f1485b5a9b44