Update patch set 2
Patch Set 2: Code-Review+1

Thanks for taking the changes, looking good.

One comment about the first comment about the property name "padBottom." Using that name >because it affects CodeMirror< is violative of the principle of containment. For example, if in the future for some reason we decided to switch from CodeMirror to Monaco (or some other editor that doesn't have a matching API), that would be kind of silly then, wouldn't it?  

Remember, your API is *your* public contract with the outside world. It can mirror an underlying API as an optimization, but it should also make sense in the absence of knowing what's happening under the hood.

Nonetheless, I don't find it *so* offensive I'm going to block. Thanks!

Patch-set: 2
Reviewer: Gerrit User 1329477 <1329477@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1