Update patch set 6
Patch Set 6:

(1 comment)

I like this, being able to edit those cookie properties is nice.
The checkboxes are a little funny, but they seem to work pretty well.
For SameSite, could we have a dropdown or something with the possible values? I think that once it goes over the protocol, it has to be one of the specific enum values defined in the protocol anyway.

Patch-set: 6
Reviewer: Gerrit User 1303398 <1303398@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>