Update patch set 7
Patch Set 7:

(1 comment)

@Brandon:

I included your change proposed in (1) but (2) is not easily changeable because of the way the CookieTable is currently implemented. After changing a single field, the cookie written back to the browser and then the new list of cookies is fetched and displayed. Depending on the timing, this leads to the editor losing the focus...

WDYT about the changes in regards to (1)?

Patch-set: 7
Reviewer: Gerrit User 1214214 <1214214@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>