Animations: Scale timeline when moving slider with keyboard
This fixes a bug where the Animations drawer timeline doesn't adjust
its scale when the keyboard is used to lengthen an animation (but does
when the mouse is used). I broke the resizing logic from the mouse
handler into its own method so that the keyboard handler can take
advantage of it as well.

Video: https://i.imgur.com/ZEImHNN.mp4

Bug: 1042102
Change-Id: I571e400106a00d334a5f0becf25898d6af9878da