WebAudio: Fix a11y web audio truncation at 200% zoom
Problem: The web audio container contents are truncated when the view
port is reduced, such as at 200% zoom.

Solution: Support scrolling if the content overflows. A container
was added to wrap the content of the web audio view so the remains
fixed at the top.

This meets WCAG 1.4.4 Resize Text success criteria
https://www.w3.org/WAI/WCAG21/quickref/#resize-text

Bug: 963183
Change-Id: I876040305a3e130b676916936079f308a9111af0