Update patch set 10
Patch Set 10:

> Patch Set 10: Code-Review-1
> 
> At very high zoom, the top toolbar underneath the tab strip has a visual glitch.

Does the following screenshot have the visual glitch you are referring to?
https://imgur.com/rCNXc6Q

The above screenshot has a visual glitch, I believe it is the one you are referring to and thank you for calling it out because it is a poor experience for users and we should fix it.

The glitch in the above screenshot existed before the change in this CR (the screenshot is from Chrome Canary 81.0.4028.0) and this change does not make the glitch worse so I recommend we get this CR merged and look into fixing the visual glitch in a separate change.

Patch-set: 10
Reviewer: Gerrit User 1341668 <1341668@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0