[css-pseudo] Handle nested pseudo-elements in devtools tree (front end)
r731964 added support for nested pseudo-elements. When a ::before or
::master that contains a nested ::marker is inspected in the devtools
tree, currently it looks like this:

  ▾ <div>
    ▾ ::before
        ::marker
        ::before
    </div>

It seems there is also a nested ::before, but this is just a poor
attempt to write a closing tag. But pseudo-elements don't have closing
tags, so this patch removes it so that the tree looks like:

  ▾ <div>
    ▾ ::before
        ::marker
    </div>

Screenshot: https://imgur.com/a/s7SYnbO

Bug: 457718
Change-Id: Id07e912f2b4830d786687d82e8dab5b2e23fdfd3