[EXPERIMENTAL] Introduce front_end/common/BUILD.gn for TypeScript type-checking
This adds a `BUILD.gn` for the `common/` folder to allow type-checking
with TypeScript. The build configuration has been fixed such that
type-checking happens on both DevTools and Chromium with the correct
paths.

Most of the files in `common/` fail to be tested with TypeScript, as
TypeScript does not support `@interface`. Trie.js was the easiest and
smallest file I could find that (with minimal changes) would pass the
TypeScript checker.

Bug: 1011811
Change-Id: I482761bfbcbf459478dc7f51f2df6576ac872de6
DISABLE_THIRD_PARTY_CHECK=TypeScript build configuration changes