Update patch set 3
Patch Set 3: Commit-Queue+1

(1 comment)

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > (1 comment)
> > 
> > +mmenke +aarontag do either of you happen to know if the net stack saves request headers in the HTTP cache, and if it would be possible to emit a Network.requestWillBeSentExtraInfo event for DevTools when we hit the HTTP cache along with the Network.responseReceivedExtraInfo event we are already getting in this case, which implies that we are already pulling out detailed response headers from the cache?
> 
> We need to store the original request headers to deal with varies response headers, though the headers we store aren't the actual headers that were actually sent (They won't have things added further down the network stack, like Connection: Keep-Alive, proxy headers, HTTP auth headers, the host header, transfer-encoding, content-length, cache-control headers, etc).
> 
> [+morlovich], who's much more familiar with the cache than I.

I see... that does at least sound a lot better than the 1-3 headers we get from blink. It may be worth looking into, and I think it would also be helpful to emit a Network.requestWillBeSentExtraInfo in this situation since we are already emitting a Network.responseReceivedExtraInfo in this case, because the frontend kinda assumes that if we get one of the two we will get both.

Patch-set: 3
Reviewer: Gerrit User 1303398 <1303398@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1
Work-in-progress: false