Update patch set 7
Patch Set 7:

(4 comments)

Hi Meredith and Alice,

thanks for sending the CL! Overall it looks good, but I have a couple of questions as I'm new to this part of the code-base as well.

Also: Would it make sense to add some tests for this? I've tried to find a layout test that exercises the code this CL touches, but wasn't able to (low test coverage is a known problem for us).

I'm adding Mathias as a reviewer as well as he is probably more knowledgeable in this area.

Sigurd

Patch-set: 7
Reviewer: Gerrit User 1214214 <1214214@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1119600 <1119600@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>