Update Chromium builds for tests
The Blink web tests for an active CL [1] are currently failing on the
bots, because the CL depends on two upstream bugfixes ([2], [3]) that
aren't yet available in the older binaries. The last of these fixes
rolled into Chromium in [4].

This patch updates to more recent Chromium binaries so that the test
outcomes may reflect reality.

[1]: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2014921
[2]: https://chromium-review.googlesource.com/c/deps/inspector_protocol/+/2015635
[3]: https://chromium-review.googlesource.com/c/deps/inspector_protocol/+/2018232
[4]: https://chromium-review.googlesource.com/c/chromium/src/+/2018175

Bug: chromium:848823, chromium:1042269, chromium:1043142
Change-Id: I02169c0c6588bbc2e53ef782913e756185066342