Add a new scope chain view for showing wasm variables.
When debugging wasm, the Scope view currently shows the engine state at
the stopped call frame. That is, the developer can inspect for instance
wasm globals, locals, and linear memory contents. When debug symbols are
available, this information is less interesting and information about
variables from the source language should be shown instead.

This CL adds a second, experiment-guarded scope view that will be
populated by the DWARF debugger language plugin. Currently, the view
does not show variable contents since that depends on larger backend
changes part of coming CLs.

Tests will follow as soon as the debugger is mockable.

Bug: chromium:1042636
Change-Id: Id562fb88bacaa958b0122cd4a9f7d948a8beb6d1